Rename `ext` to `commands`
